@@ -1,7 +1,42 @@
 const express = require('express');
+const nodemailer = require('nodemailer');
 const router = express.Router();
-const { submitContact } = require('../controllers/contactController');
 
-router.post('/', submitContact);
+router.post('/', async (req, res) => {
+  const { name, email, message } = req.body;
+
+  // Basic validation
+  if (!name || !email || !message) {
+    return res.status(400).json({ error: 'All fields are required.' });
+  }
+
+  try {
+    // Create transporter
+    const transporter = nodemailer.createTransport({
+      service: 'gmail',
+      auth: {
+        user: process.env.EMAIL_USER,
+        pass: process.env.EMAIL_PASS,
+      },
+    });
+
+    // Send email
+    await transporter.sendMail({
+      from: `"${name}" <${email}>`,
+      to: process.env.EMAIL_TO,
+      subject: 'Portfolio Contact Form Submission',
+      text: message,
+      html: `<p><strong>Name:</strong> ${name}</p>
+             <p><strong>Email:</strong> ${email}</p>
+             <p><strong>Message:</strong><br/>${message}</p>`,
+    });
+
+    res.json({ success: true, message: 'Email sent successfully!' });
+    console.log('Email sent successfully');
+  } catch (err) {
+    console.error('Error sending email:', err);
+    res.status(500).json({ error: 'Failed to send email. Try again later.' });
+  }
+});
 
 module.exports = router;
